+++ src/components/customs/Signup.tsx
@@ -10,23 +10,17 @@ import { useRouter } from "next-intl/client";
 import Image from "next/image";
 import { Controller, SubmitHandler, useForm } from "react-hook-form";
 import { toast } from "react-toastify";
-
-type SignUpRequest = Omit<SignUpInputs, "confirmPassword">; // Omit confirmPassword from SignUpInputs to create SignUpRequest
-type SignUpInputs = {
-  name: string;
-  username: string;
-  email: string;
-  signupPassword: string;
-  confirmPassword: string;
-};
-type LoginInputs = {
-  usernameOrEmail: string;
-  password: string;
-};
-
-type ForgotPassword = {
-  forgotPasswordEmail: string;
-};
+import {
+  ForgotPassword,
+  LoginInputs,
+  SignUpInputs,
+  SignUpRequest,
+} from "@/types";
+import dynamic from "next/dynamic";
+const SignupForm = dynamic(
+  () => import("@/components/customs/Signup/SignupForm"),
+  { ssr: false }
+);
 
 export default function Signup({
   params,
@@ -108,163 +102,17 @@ export default function Signup({
   return (
     <div className="absolute grid place-items-center w-full h-[calc(100%-64px)] p-2">
       {userParam && userParam === "new" ? (
-        <form
-          onSubmit={handleSubmit(onSignupSubmit)}
-          className="flex flex-col gap-2 w-11/12 sm:w-full max-w-xl bg-content1 backdrop-saturate-150 p-6 sm:p-12 rounded-xl"
-        >
-          <Button
-            variant="bordered"
-            color="primary"
-            onClick={() => onProviderSignin("google")}
-            onKeyDown={(e) =>
-              onEnterAndSpace(e, () => onProviderSignin("google"))
-            }
-            startContent={
-              <Image
-                src={"/svg/providers/google.svg"}
-                width={24}
-                height={24}
-                alt="google-icon"
-              />
-            }
-          >
-            {t("Sign up with Google")}
-          </Button>
-          <Divider></Divider>
-          <div>
-            <h1 className="text-xl sm:text-2xl lg:text-4xl font-bold">
-              {t("Create a new account")}
-            </h1>
-          </div>
-          <Controller
-            name="name"
-            rules={{
-              required: true,
-              onChange: (e) => {
-                if (e.target.value.length > 0) {
-                  clearErrors("name");
-                }
-              },
-            }}
-            control={control}
-            render={({ field, fieldState: { error } }) => (
-              <Input
-                {...field}
-                label={t("Name")}
-                isRequired
-                color="primary"
-                variant="underlined"
-                errorMessage={errors.name?.message}
-                isInvalid={error !== undefined}
-              />
-            )}
-          />
-          <Controller
-            name="username"
-            rules={{
-              required: true,
-              onChange: (e) => {
-                if (e.target.value.length > 0) {
-                  clearErrors("username");
-                }
-              },
-            }}
-            control={control}
-            render={({ field, fieldState: { error } }) => (
-              <Input
-                {...field}
-                label={t("Username")}
-                isRequired
-                color="primary"
-                variant="underlined"
-                errorMessage={errors.username?.message}
-                isInvalid={error !== undefined}
-              />
-            )}
-          />
-          <Controller
-            control={control}
-            name="email"
-            rules={{
-              required: true,
-              pattern: {
-                value: /^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/g,
-                message: "Please enter a valid email",
-              },
-            }}
-            render={({ field, fieldState: { error } }) => (
-              <Input
-                type="email"
-                {...field}
-                label={t("Email")}
-                isRequired
-                color="primary"
-                variant="underlined"
-                errorMessage={errors.email?.message}
-                isInvalid={error !== undefined}
-              />
-            )}
-          />
-          <Controller
-            control={control}
-            name="signupPassword"
-            rules={{
-              required: true,
-              minLength: {
-                value: 8,
-                message: "Password must have at least 8 characters",
-              },
-            }}
-            render={({ field, fieldState: { error } }) => (
-              <Input
-                {...field}
-                label={t("Password")}
-                isRequired
-                color="primary"
-                variant="underlined"
-                errorMessage={errors.signupPassword?.message}
-                isInvalid={error !== undefined}
-                type="password"
-              />
-            )}
-          />
-          <Controller
-            control={control}
-            name="confirmPassword"
-            rules={{
-              required: true,
-              validate: (value) =>
-                value === watch("signupPassword") || "Passwords do not match",
-            }}
-            render={({ field, fieldState: { error } }) => (
-              <Input
-                {...field}
-                label={t("Confirm Password")}
-                isRequired
-                color="primary"
-                variant="underlined"
-                errorMessage={errors.confirmPassword?.message}
-                isInvalid={error !== undefined}
-                type="password"
-              />
-            )}
-          />
-
-          <Button color="primary" variant="ghost" type="submit">
-            {t("Sign Up Button")}
-          </Button>
-          <p>
-            {t("Already have an account?")}{" "}
-            <Link
-              href={`${
-                params.callbackUrl ? `?callbackUrl=${params.callbackUrl}` : ""
-              }`}
-              className="underline hover:text-primary transition-colors focus-visible:outline-none focus-visible:text-primary"
-            >
-              {t("Login")}
-            </Link>
-          </p>
-        </form>
+        <SignupForm
+          control={control}
+          handleSubmit={handleSubmit}
+          onProviderSignin={onProviderSignin}
+          onSignupSubmit={onSignupSubmit}
+          clearErrors={clearErrors}
+          watch={watch}
+          errors={errors}
+          t={t}
+          params={params}
+        />
       ) : userParam === "forgot" ? (
         <>
           <form
